+++ opt-env-conf/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * `xdgYamlConfigFile` now returns a `Path Abs File` instead of a `FilePath`.
 * Fixed a bug in `withFirstYamlConfig` and `withCombinedYamlConfig` in which the `--config-file` option was required.
+* `--run-settings-check` now allows you to define a static settings check.
 
 ## [0.0.0.0] - 2024-07-08